CASMCMS-7229: Include calling file and line number in log entries
